Cleanup twxvu bindings to assume installed libtwxvu.so
